+++ src/types/JourneyOutput.ts
@@ -1,3 +1,5 @@
+import { TrainOperator, Ticket, StationInfo } from './JourneyResponse';
+
 export type JourneyOutput = {
   journeyID: string;
   departureTime: string;
@@ -6,19 +8,4 @@ export type JourneyOutput = {
   tickets: Array<Ticket>;
   originStation: StationInfo;
   destinationStation: StationInfo;
-};
-
-export type StationInfo = {
-  displayName: string;
-  crs: string;
-};
-
-export type TrainOperator = {
-  code: string;
-  name: string;
-};
-
-export type Ticket = {
-  name: string;
-  priceInPennies: number;
 };
